+++ src/curve/sec2/mod.rs
@@ -22,6 +22,8 @@ pub mod p224r1;
 pub mod p256k1;
 #[cfg(feature = "p256r1")]
 pub mod p256r1;
+#[cfg(feature = "p256r1")]
+pub mod p256r1_mul;
 #[cfg(feature = "p384r1")]
 pub mod p384r1;
 #[cfg(feature = "p521r1")]